1.修改指令定义 2.实现二元运算符opcode生成和打印代码

+++ Source/QLanguage/Parser/SyntaxTree/SyntaxTree_Base.h
@@ -107,7 +107,7 @@ class SyntaxTree_Do;
 
         inline virtual const bool isConstValue()const { return false; }
         inline virtual const bool toBool()const { return false; }
-        inline virtual const ushort getBlock()const { return -1; }
+        inline virtual const uchar  getBlock()const { return -1; }
         inline virtual const ushort getIndex()const { return -1; }
 
         virtual const bool operator==(const SyntaxTree_Base& x)const=0;
